popups deleted move to pages
