+++ app/[lang]/arbswap/layout.tsx
@@ -8,14 +8,8 @@ export const metadata: Metadata = {
 
 type RootLayoutProps = Readonly<{
   children: React.ReactNode
-  auth: React.ReactNode
 }>
 
-export default function DashboardLayout({ children, auth }: RootLayoutProps) {
-  return (
-    <section>
-      {children}
-      {auth}
-    </section>
-  )
+export default function DashboardLayout({ children }: RootLayoutProps) {
+  return <section>{children}</section>
 }
