+++ components/HomePage/Arbswap/Auth/Layout.tsx
@@ -22,7 +22,7 @@ const Layout = ({ children }: LayoutProp) => {
         className="-z-10"
         quality={95}
       />
-      <div className="ml-3 mt-32 flex justify-center">
+      <div className="flex justify-center md:mt-0 lg:mt-6 xl:mt-12 2xl:mt-28">
         <Image
           src="/img/nexarb_layer1.png"
           alt="nexarb_layer1"
@@ -30,7 +30,7 @@ const Layout = ({ children }: LayoutProp) => {
           height={100}
         />
       </div>
-      <div className="mt-12 h-[495px] w-[449px] rounded-[38px] border border-slate-700 bg-transparent bg-opacity-70 p-9 backdrop-blur-[100px] mx-auto">
+      <div className="mx-auto max-w-md rounded-[38px] border border-slate-700 bg-transparent bg-opacity-70 p-9 backdrop-blur-[100px] lg:mt-6 xl:mt-12">
         {children}
       </div>
     </div>
